connect to the debugger
