+++ ddtrace/tracer.py
@@ -54,7 +54,6 @@ from .internal.writer import AgentWriter
 from .internal.writer import LogWriter
 from .internal.writer import TraceWriter
 from .provider import DefaultContextProvider
-from .remoteconfig._worker import RemoteConfigWorker
 from .sampler import BasePrioritySampler
 from .sampler import BaseSampler
 from .sampler import DatadogSampler
