+++ riotfile.py
@@ -314,15 +314,6 @@ venv = Venv(
             },
         ),
         Venv(
-            name="remoteconfig",
-            command="pytest {cmdargs} tests/remoteconfig/",
-            pkgs={
-                "msgpack": latest,
-                "httpretty": "==0.9.7",
-            },
-            pys=select_pys(),
-        ),
-        Venv(
             name="debugger",
             command="pytest {cmdargs} tests/debugging/",
             pkgs={
