--- tests/remoteconfig/payloads/base-req-1.json
@@ -1,1 +1,0 @@
-{"client":{"state":{"targets_version":1}}}
