--- tests/remoteconfig/payloads/base-req-2.json
@@ -1,1 +1,0 @@
-{"client":{"state":{"targets_version":2,"config_states":[{"id":"asm1","version":1,"product":"ASM"}]}}}
