--- tests/remoteconfig/payloads/base-res-1.json
@@ -1,1 +1,0 @@
-{"targets":"eyJzaWduZWQiOnsiX3R5cGUiOiJ0YXJnZXRzIiwic3BlY192ZXJzaW9uIjoiMS4wIiwidmVyc2lvbiI6MiwiZXhwaXJlcyI6IjIwMjItMDktMTdUMTY6MTI6MDBaIiwidGFyZ2V0cyI6eyJkYXRhZG9nLzIvQVNNL2FzbTEvY29uZmlnIjp7Imxlbmd0aCI6NTYsImhhc2hlcyI6eyJzaGEyNTYiOiI4OTEyNTU1YmYxY2FlYWExZmY5MThiMmRlZTNiZjMwOTMwMDNmMjdjNDJhZmM1YzZiMDFiZGZiMzgwZGNiYjlhIn0sImN1c3RvbSI6eyJ2IjoxfX19fSwic2lnbmF0dXJlcyI6W3sia2V5aWQiOiJlZDc2NzJjOWEyNGFiZGE3ODg3MmVlMzJlZTcxYzdjYjFkNTIzNWU4ZGI0ZWNiZjFjYTI4YjljNTBlYjc1ZDllIiwic2lnIjoiODc2ZGMyZTZmMjgyMzAyZjFiODlhYzA0ODkyM2FlMDc1NmEyOTM1MGY3ZmNiMGM5NjdkYzUxNzlkM2Q0MzdjYTNjZmU3OTkyYmIxZGE2YmY3MGU2YjhhNjU3MTJiODQ4OWIxMzE3ZThjOTAwODcwMmUzMzkzZWQ0NDBjNjU1MDQifV19","target_files":[{"path":"datadog/2/ASM/asm1/config","raw":"eyJydWxlc19vdmVycmlkZSI6W3siaWQiOiJjcnMtMDQyLTAwMSIsImVuYWJsZWQiOnRydWV9XX0="}],"client_configs":["datadog/2/ASM/asm1/config"]}
