--- tests/remoteconfig/payloads/base-res-2.json
@@ -1,1 +1,0 @@
-{}
